backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
passlib==1.7.4
email-validator==2.1.0.post1
pydantic-settings==2.1.0
python-jose==3.3.0